Provide more user settings (#26)
* Provide actual hide sidebar/menu/etc. settings for ST3

I think these would be better than toggling to the opposite
of what the main window is - at least for me my main window
is not consistent with regard to sidebar.

* Provide a 'display_prefix' setting to let the user specify a title prefix

With seveeral open windows, it's hard to figure out which sublime window
has the diff info. This change adds a 'display_prefix' setting that
lets the user specify a string to prefix to the tab and window titles,
to help distinguish the diff window from the others.