Fix 'Could not load colour scheme' error message
At least in MacOS, the default color_scheme Preference is set like this:
  "color_scheme": "Monokai.sublime-color-scheme"

It lacks of the 'Packages/Color Scheme - Default/' package path. Given
that, when load_resource() is called it can't load it and falls back to
a blank color scheme.

The provided solution prepends the said package path to the value
obtained from the settings when it does not start with 'Packages/'.

See also related issue #31