Redo actual diffing
The code did find all the actual inline-modification.  It only detected
what `ndiff` marks with a "?"; however, only very simple modification
are marked as such.

It did the overall diff and the costly diffs per line in one go.  It is
beneficial to entangle this to give the user a faster answer.  The
inline diffs are optional, or: lazy enhancement, anyway.

The line statistics were off.  The intra-line modifications were not
subtracted from the line removals and additions.

Fixes #7